V2.3 Mar/2/2022
Whats New main reason for this being a new release is changing the color scheme for the css its now a terquise, redish-orange and white with lots of box shadow. also added css to the classes pages but right now its just a copy of the css for indexV2.html  linked to all the classes I am going to use it in the futear to globly style those pages but for now its just a place holder. what else I added the info for classes 7, 8 and 9. and thats everything new I think.